+++ Runtime/Integrations/Zenject/RelationalSceneLoadListener.cs
@@ -10,6 +10,7 @@ namespace WallstopStudios.UnityHelpers.Integrations.Zenject
     using UnityEngine;
     using UnityEngine.SceneManagement;
     using WallstopStudios.UnityHelpers.Core.Attributes;
+    using WallstopStudios.UnityHelpers.Core.Extension;
     using WallstopStudios.UnityHelpers.Tags;
     using WallstopStudios.UnityHelpers.Utils;
 
@@ -65,9 +66,9 @@ namespace WallstopStudios.UnityHelpers.Integrations.Zenject
             {
                 // Fallback: scan all components in new scene and assign when type has relational fields
                 bool includeInactiveAll = _options.IncludeInactive;
-                Component[] all = includeInactiveAll
-                    ? UnityEngine.Object.FindObjectsOfType<Component>(true)
-                    : UnityEngine.Object.FindObjectsOfType<Component>(false);
+                Component[] all = UnityObjectExtensions.FindObjectsOfTypeShim<Component>(
+                    includeInactiveAll
+                );
 
                 for (int i = 0; i < all.Length; i++)
                 {
@@ -196,9 +197,9 @@ namespace WallstopStudios.UnityHelpers.Integrations.Zenject
                 }
             }
 
-            Component[] all = includeInactive
-                ? UnityEngine.Object.FindObjectsOfType<Component>(true)
-                : UnityEngine.Object.FindObjectsOfType<Component>(false);
+            Component[] all = UnityObjectExtensions.FindObjectsOfTypeShim<Component>(
+                includeInactive
+            );
 
             for (int i = 0; i < all.Length; i++)
             {